+++ src/ballet/bn254/fd_bn254_scalar.h
@@ -12,6 +12,7 @@
 
 #include "../fd_ballet_base.h"
 #include "../bigint/fd_uint256.h"
+#include <stdint.h>
 #include "../../third_party/fiat-crypto/bn254_scalar_64.c"
 
 /* The implementation is based on fiat-crypto.
@@ -48,14 +49,14 @@ fd_bn254_scalar_validate( fd_bn254_scalar_t const * s ) {
 static inline fd_bn254_scalar_t *
 fd_bn254_scalar_from_mont( fd_bn254_scalar_t *       r,
                            fd_bn254_scalar_t const * a ) {
-  fiat_bn254_scalar_from_montgomery( (uint64_t *)r->limbs, (uint64_t const *)a->limbs );
+  fiat_bn254_scalar_from_montgomery( (uint64_t *)(uintptr_t)r->limbs, (uint64_t const *)(uintptr_t)a->limbs );
   return r;
 }
 
 static inline fd_bn254_scalar_t *
 fd_bn254_scalar_to_mont( fd_bn254_scalar_t *       r,
                          fd_bn254_scalar_t const * a ) {
-  fiat_bn254_scalar_to_montgomery( (uint64_t *)r->limbs, (uint64_t const *)a->limbs );
+  fiat_bn254_scalar_to_montgomery( (uint64_t *)(uintptr_t)r->limbs, (uint64_t const *)(uintptr_t)a->limbs );
   return r;
 }
 
@@ -63,7 +64,7 @@ static inline fd_bn254_scalar_t *
 fd_bn254_scalar_add( fd_bn254_scalar_t *       r,
                      fd_bn254_scalar_t const * a,
                      fd_bn254_scalar_t const * b ) {
-  fiat_bn254_scalar_add( (uint64_t *)r->limbs, (uint64_t const *)a->limbs, (uint64_t const *)b->limbs );
+  fiat_bn254_scalar_add( (uint64_t *)(uintptr_t)r->limbs, (uint64_t const *)(uintptr_t)a->limbs, (uint64_t const *)(uintptr_t)b->limbs );
   return r;
 }
 
@@ -73,14 +74,14 @@ static inline fd_bn254_scalar_t *
 fd_bn254_scalar_mul( fd_bn254_scalar_t *       r,
                      fd_bn254_scalar_t const * a,
                      fd_bn254_scalar_t const * b ) {
-  fiat_bn254_scalar_mul( (uint64_t *)r->limbs, (uint64_t const *)a->limbs, (uint64_t const *)b->limbs );
+  fiat_bn254_scalar_mul( (uint64_t *)(uintptr_t)r->limbs, (uint64_t const *)(uintptr_t)a->limbs, (uint64_t const *)(uintptr_t)b->limbs );
   return r;
 }
 
 static inline fd_bn254_scalar_t *
 fd_bn254_scalar_sqr( fd_bn254_scalar_t *       r,
                      fd_bn254_scalar_t const * a ) {
-  fiat_bn254_scalar_square( (uint64_t *)r->limbs, (uint64_t const *)a->limbs );
+  fiat_bn254_scalar_square( (uint64_t *)(uintptr_t)r->limbs, (uint64_t const *)(uintptr_t)a->limbs );
   return r;
 }
